Refactor sql dump script to use variables
This commit updates the csv_dump.sql script to use variables for the table and database names. This will allow for more flexibility and reusability of the script since the table and database names can now easily be changed. Manually repeating table names has been removed to reduce potential errors.